frontend/requirements.txt
streamlit==1.24.1
requests==2.31.0
Pillow==10.0.0
